requirements.txt
torch
torchvision
numpy
opencv-python
PyYAML
tqdm
ultralytics==8.0.20